Changes to PID library to allow it to be configured to not be time based.
- Allow SampleTime to be set to 0 to force Compute() to run every time
it's called. In this case, the PID algorithm is not time based, and the
user should not call it every time through loop(), but rather only based
on some event (e.g. timer, button press, external input).

- Add GetITerm() and GetLastInput() functions to return the current
ITerm and LastInput, respectively.  Being able to get and store these
values allows the user of the PID library to reinitialize the PID
library and continue using the same PID curve and calculation across
resets.  (Pass the saved ITerm as output and lastInput as input to the
PID constructor to reinitialize and pick up where you left off.)